move log sender test outside of docker container
